# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deep Learning CNN+LSTM API",
    "description": "API REST pour le projet de Deep Learning avec CNN+LSTM",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@deeplearning.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
...
  "paths": {
...
    "/api/training/start": {
      "post": {
        "tags": ["Training"],
        "summary": "Démarrer l'entraînement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "epochs": {"type": "integer", "example": 10},
                  "batchSize": {"type": "integer", "example": 32},
                  "learningRate": {"type": "number", "example": 0.001},
                  "validationSplit": {"type": "number", "example": 0.2}
                },
                "required": ["epochs", "batchSize", "learningRate"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entraînement démarré"
          }
        }
      }
    },
...
    "/api/training/evaluate": {
      "post": {
        "tags": ["Training"],
        "summary": "Évaluer le modèle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataset": {
                    "type": "string",
                    "enum": ["test", "validation"],
                    "example": "test"
                  }
                },
                "required": ["dataset"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Résultats d'évaluation"
          }
        }
      }
    },
    "/api/training/predict": {
      "post": {
        "tags": ["Training"],
        "summary": "Faire une prédiction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inputData": {
                    "type": "array",
                    "description": "Données d'entrée"
                  }
                },
                "required": ["inputData"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prédiction"
          }
        }
      }
    },
...
  },
...
}